Fix: equivs-build doesn't work if & is part of PWD
Unfortunately this is a real world situation when we are using jenkins
and mk-build-deps in the test job. We are using label expressions to
select the correct test configuration (e.g. "amd64&&precise" to tell
jenkins I want to test on amd64 and precise) . For some resons the label
expression used for selecting the jenkins node is used in the path of
the working directory. Hence this bug.